module-2/10_Review/review-final/java/src/test/java/com/techelevator/dao/JdbcReservationDaoTests.java
package com.techelevator.dao;

import com.techelevator.model.Reservation;
import org.junit.Before;
import org.junit.Test;

import java.time.LocalDate;
import java.util.List;

import static org.junit.Assert.assertEquals;

public class JdbcReservationDaoTests extends BaseDaoTests {

    private ReservationDao dao;

    @Before
    public void setup() {
        dao = new JdbcReservationDao(dataSource);
    }

    @Test
    public void createReservation_Should_ReturnNewReservationId() {
        int reservationCreated = dao.createReservation(9999,
                "TEST NAME",
                LocalDate.now().plusDays(1),
                LocalDate.now().plusDays(3));

        assertEquals(reservationCreated, 1);
    }

    @Test
    public void getUpcomingReservations_Should_ReturnReservations() {
        List<Reservation> reservations = dao.getUpcomingReservations(99);

        assertEquals(2,reservations.size());
    }

}
